+++ ldap4j-java/src/main/java/hu/gds/ldap4j/ldap/ModifyResponse.java
@@ -1,6 +1,7 @@
 package hu.gds.ldap4j.ldap;
 
 import hu.gds.ldap4j.net.ByteBuffer;
+import java.util.List;
 import java.util.Objects;
 import org.jetbrains.annotations.NotNull;
 
@@ -8,8 +9,10 @@ public record ModifyResponse(
         @NotNull LdapResult ldapResult) {
     public static class Reader implements MessageReader<ModifyResponse> {
         @Override
-        public void check(@NotNull ModifyResponse message) throws Throwable {
-            message.ldapResult().checkSuccess();
+        public void check(
+                @NotNull List<@NotNull Control> controls, @NotNull ModifyResponse message, int messageId)
+                throws Throwable {
+            message.ldapResult().checkSuccess(controls, messageId);
         }
 
         @Override